[feat] principal 인증 기능 추가

+++ src/main/java/com/example/smart_edu_team/user/CommonUserController.java
@@ -7,6 +7,8 @@ import org.springframework.ui.Model;
 import org.springframework.validation.BindingResult;
 import org.springframework.web.bind.annotation.*;
 
+import java.security.Principal;
+import java.util.Objects;
 import java.util.Optional;
 
 @Controller
@@ -18,10 +20,10 @@ public class CommonUserController implements UserController {
 
 
     @Override
-    @GetMapping("/{username}")
-    public String getUserByUsername(@PathVariable String username, Model model) {
+    public String getUserByUsername(@PathVariable String username, Model model, Principal principal) {
         Optional<UserDTO> user = userService.findByUsername(username);
-        if (user.isPresent()) {
+        if (user.isPresent() &&
+                Objects.equals(principal.getName(), username)) {
             model.addAttribute("user", user.get());
             return "user/details";
         }
@@ -29,14 +31,12 @@ public class CommonUserController implements UserController {
     }
 
     @Override
-    @GetMapping("/signup")
     public String showSignupForm(Model model) {
         model.addAttribute("user", new UserDTO());
         return "user/signup";
     }
 
     @Override
-    @PostMapping("/signup")
     public String signup(@ModelAttribute UserDTO UserDTO, BindingResult bindingResult) {
         if (userService.findByUsername(UserDTO.getUsername()).isPresent()) {
             bindingResult.reject("signupFailed", "이미 등록된 사용자입니다.");
@@ -47,23 +47,16 @@ public class CommonUserController implements UserController {
     }
 
     @Override
-    @GetMapping("/login")
     public String showLoginForm(Model model) {
         model.addAttribute("user", new UserDTO());
         return "user/login";
     }
 
-    /**
-     *
-     * @param username
-     * @param model
-     * @return 유저 수정 템플릿을 리턴하나 mypage가 아닐 시, not_found를 리턴합니다..
-     */
     @Override
-    @GetMapping("/edit/{username}")
-    public String showEditForm(@PathVariable String username, Model model) {
+    public String showEditForm(@PathVariable String username, Model model, Principal principal) {
         Optional<UserDTO> user = userService.findByUsername(username);
-        if (user.isPresent()) {
+        if (user.isPresent() &&
+                Objects.equals(principal.getName(), username)) {
             model.addAttribute("user", user.get());
             return "user/edit";
         }
@@ -71,16 +64,22 @@ public class CommonUserController implements UserController {
     }
 
     @Override
-    @PostMapping("/edit/{username}")
-    public String updateUser(@PathVariable String username, @ModelAttribute UserDTO userDetails) {
-        userService.updateUser(username, userDetails);
-        return "redirect:/users";
+    public String updateUser(@PathVariable String username, @ModelAttribute UserDTO userDetails, Principal principal) {
+        if ( Objects.equals(principal.getName(), username)) {
+            userService.updateUser(username, userDetails);
+            return "redirect:/users";
+        } else {
+            return "user/not_found";
+        }
     }
 
     @Override
-    @PostMapping("/delete/{username}")
-    public String deleteUser(@PathVariable String username) {
-        userService.deleteUser(username);
-        return "redirect:/users";
+    public String deleteUser(@PathVariable String username, Principal principal) {
+        if ( Objects.equals(principal.getName(), username)) {
+            userService.deleteUser(username);
+            return "redirect:/users";
+        } else {
+            return "user/not_found";
+        }
     }
 }
